+++ src/builtins/history.rs
@@ -1,15 +1,45 @@
 use std::io::Write;
 use std::path::Path;
 
+use chrono::NaiveDateTime;
 use rusqlite::Connection as Conn;
 use rusqlite::NO_PARAMS;
+use structopt::StructOpt;
 
 use crate::history;
+use crate::parsers;
+use crate::shell;
 use crate::types;
 
-pub fn run(cmd: &types::Command) -> i32 {
-    let tokens = &cmd.tokens;
+#[derive(Debug, StructOpt)]
+#[structopt(name = "history", about = "History in cicada shell")]
+struct OptMain {
+    #[structopt(short, long, help = "For current session only")]
+    session: bool,
 
+    #[structopt(short, long, help = "Search old items first")]
+    asc: bool,
+
+    #[structopt(short, long, help = "For current directory only")]
+    pwd: bool,
+
+    #[structopt(short, long, help = "Only show ROWID")]
+    only_id: bool,
+
+    #[structopt(short, long, help = "Do not show ROWID")]
+    no_id: bool,
+
+    #[structopt(short="d", long, help = "Show date")]
+    show_date: bool,
+
+    #[structopt(short, long, default_value = "20")]
+    limit: i32,
+
+    #[structopt(name = "PATTERN", default_value = "", help = "You can use % to match anything")]
+    pattern: String,
+}
+
+pub fn run(sh: &shell::Shell, cmd: &types::Command) -> i32 {
     let hfile = history::get_history_file();
     let path = Path::new(hfile.as_str());
     if !path.exists() {
@@ -24,35 +54,49 @@ pub fn run(cmd: &types::Command) -> i32 {
         }
     };
 
-    if tokens.len() == 1 {
-        return list_current_history(&conn);
-    } else if tokens.len() == 2 {
-        search_history(&conn, &tokens[1].1);
-    } else if tokens.len() == 3 {
-        if tokens[1].1 != "delete" {
-            println_stderr!("only support: history delete");
+    let tokens = &cmd.tokens;
+    let args = parsers::parser_line::tokens_to_args(tokens);
+
+    if args.len() >= 2 && args[1] == "delete" {
+        if args.len() != 3 {
+            println_stderr!("USAGE: history delete <row-id>");
             return 1;
         }
 
         if let Ok(rowid) = tokens[2].1.parse::<usize>() {
             delete_history_item(&conn, rowid);
+            return 0;
         } else {
             println_stderr!("history delete: a row number is needed");
             return 1;
         }
-    } else {
-        println_stderr!("history: only take one or two args");
-        return 1;
     }
-    0
+
+    let opt = OptMain::from_iter(args);
+    return list_current_history(sh, &conn, &opt);
 }
 
-fn list_current_history(conn: &Conn) -> i32 {
+fn list_current_history(sh: &shell::Shell, conn: &Conn, opt: &OptMain) -> i32 {
     let history_table = history::get_history_table();
-    let sql = format!(
-        "SELECT rowid, inp FROM {} ORDER BY tsb desc limit 20;",
-        history_table
-    );
+    let mut sql = format!("SELECT ROWID, inp, tsb FROM {} WHERE ROWID > 0",
+                          history_table);
+    if opt.pattern.len() > 0 {
+        sql = format!("{} AND inp LIKE '%{}%'", sql, opt.pattern)
+    }
+    if opt.session {
+        sql = format!("{} AND sessionid = '{}'", sql, sh.session_id)
+    }
+    if opt.pwd {
+        sql = format!("{} AND info like '%dir:{}|%'", sql, sh.current_dir)
+    }
+
+    if opt.asc {
+        sql = format!("{} ORDER BY tsb", sql);
+    } else {
+        sql = format!("{} order by tsb desc", sql);
+    };
+    sql = format!("{} limit {} ", sql, opt.limit);
+
     let mut stmt = match conn.prepare(&sql) {
         Ok(x) => x,
         Err(e) => {
@@ -87,7 +131,24 @@ fn list_current_history(conn: &Conn) -> i32 {
                             return 1;
                         }
                     };
-                    println!("{}: {}", row_id, inp);
+
+                    if opt.no_id {
+                        println!("{}", inp);
+                    } else if opt.only_id {
+                        println!("{}", row_id);
+                    } else if opt.show_date {
+                        let tsb: f64 = match row.get(2) {
+                            Ok(x) => x,
+                            Err(e) => {
+                                println_stderr!("history: error: {:?}", e);
+                                return 1;
+                            }
+                        };
+                        let dt = NaiveDateTime::from_timestamp(tsb as i64, 0);
+                        println!("{}: {}: {}", row_id, dt.date(), inp);
+                    } else {
+                        println!("{}: {}", row_id, inp);
+                    }
                 } else {
                     return 0;
                 }
@@ -95,61 +156,6 @@ fn list_current_history(conn: &Conn) -> i32 {
             Err(e) => {
                 println_stderr!("history: rows next error: {:?}", e);
                 return 1;
-            }
-        }
-    }
-}
-
-fn search_history(conn: &Conn, q: &str) {
-    let history_table = history::get_history_table();
-    let sql = format!(
-        "SELECT ROWID, inp FROM {}
-             WHERE inp like '%{}%'
-             ORDER BY tsb desc limit 50;",
-        history_table, q
-    );
-    let mut stmt = match conn.prepare(&sql) {
-        Ok(x) => x,
-        Err(e) => {
-            println_stderr!("history: prepare select error: {:?}", e);
-            return;
-        }
-    };
-
-    let mut rows = match stmt.query(NO_PARAMS) {
-        Ok(x) => x,
-        Err(e) => {
-            println_stderr!("history: query error: {:?}", e);
-            return;
-        }
-    };
-
-    loop {
-        match rows.next() {
-            Ok(_rows) => {
-                if let Some(row) = _rows {
-                    let row_id: i32 = match row.get(0) {
-                        Ok(x) => x,
-                        Err(e) => {
-                            println_stderr!("history: error: {:?}", e);
-                            return;
-                        }
-                    };
-                    let inp: String = match row.get(1) {
-                        Ok(x) => x,
-                        Err(e) => {
-                            println_stderr!("history: error: {:?}", e);
-                            return;
-                        }
-                    };
-                    println!("{}: {}", row_id, inp);
-                } else {
-                    return;
-                }
-            }
-            Err(e) => {
-                println_stderr!("history: rows next error: {:?}", e);
-                return;
             }
         }
     }